Automatically add catches to trip #900 (#1102)

+++ mobile/test/widget/data_importer_test.dart
@@ -98,7 +98,9 @@ void main() {
   });
 
   tearDown(() {
-    tmpDir.deleteSync(recursive: true);
+    if (tmpDir.existsSync()) {
+      tmpDir.deleteSync(recursive: true);
+    }
   });
 
   DataImporter defaultImporter({
